- Moved logic from Main screen to building, from building into new file building section. - Adding new buildings, does not cause overlapping coordinates.
